software/requirements.txt
hexdump==3.3
pyserial==3.5
tqdm==4.62.2
